+++ app/task/MaaFW/tools/notify/report.py
@@ -7,11 +7,9 @@ from app.core.notify import (
     DispatchResult,
     NotifyPayload,
     dispatch,
-    global_target,
-    should_send_result,
     statistic_targets,
 )
-from app.tools.game_sign_notify import dispatch_task_report, get_task_game_sign_summary
+from app.task.notify_core import push_proxy_result
 from app.utils import get_logger
 
 logger = get_logger("MaaFW 通知工具")
@@ -41,50 +39,10 @@ async def push_notification(
     logger.info(f"开始推送通知, 模式: {mode}, 标题: {title}")
 
     if mode == "代理结果":
-        return await _push_proxy_result(title, message, task_info)
+        return await push_proxy_result(title=title, message=message, task_info=task_info)
     if mode == "统计信息":
         return await _push_statistics(title, message, user_config)
     return DispatchResult()
-
-
-async def _push_proxy_result(
-    title: str, message: dict, task_info: object | None
-) -> DispatchResult:
-    """推送脚本级「代理结果」报告（全局渠道）。"""
-
-    if not should_send_result(message):
-        return DispatchResult()
-
-    message_text = (
-        f"任务开始时间: {message['start_time']}, 结束时间: {message['end_time']}\n"
-        f"已完成数: {message['completed_count']}, "
-        f"未完成数: {message['uncompleted_count']}\n\n"
-        f"{message['result']}"
-    )
-    message_html = Config.notify_env.get_template("general_result.html").render(message)
-    counts = (
-        f"已完成用户数: {message['completed_count']}, "
-        f"未完成用户数: {message['uncompleted_count']}"
-    )
-    summary_text = (
-        get_task_game_sign_summary(task_info)
-        if task_info is not None and message.get("game_sign_summary")
-        else ""
-    )
-    return await dispatch_task_report(
-        NotifyPayload(
-            title=title,
-            text=message_text,
-            html=message_html,
-            system_title=message.get("system_title") or title.replace("报告", "已完成！"),
-            system_message=counts,
-            system_ticker=counts,
-            system_timeout=10,
-        ),
-        [global_target(include_system=True)],
-        task_info,
-        summary_text=summary_text,
-    )
 
 
 async def _push_statistics(
